+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
+### [0.2.0] ????-??-??
+
+* Added `Awesomite\Chariot\Pattern\PatternInterface` - possibility to conversions url params, e.g. date in format YYYY-mm-dd to \DateTime object
+* Added [behat] tests
+
 ### [0.1.0] 2017-07-20
     
 * Initial public release
 
 [0.1.0]: https://github.com/awesomite/chariot/tree/v0.1.0
+[behat]: http://behat.org